PREDEF PRES TO DGV

+++ SQL/predefined_prescription_related_tables.sql
@@ -11,7 +11,7 @@ CREATE TABLE predefined_prescription_dt(
 	predef_pres_id INT,
     drug_id INT,
     sub_drug_id INT,
-    dosage INT,
+    dosage DECIMAL(8, 4),
     unit INT,
     preparation_method INT
 );
